price_monitor/requirements.txt
httpx==0.27.0
pydantic==2.7.1
pyyaml==6.0.1
apscheduler==3.10.4
ollama==0.2.1
